+++ app/src/main/java/com/example/ishopping/ui/CartFragment.kt
@@ -1,6 +1,7 @@
-package com.example.ishopping
+package com.example.ishopping.ui
 
 import androidx.fragment.app.Fragment
+import com.example.ishopping.R
 
 class CartFragment: Fragment(R.layout.fragment_cart) {
 }
