+++ app/src/main/java/com/example/ishopping/ui/LikeFragment.kt
@@ -1,6 +1,7 @@
-package com.example.ishopping
+package com.example.ishopping.ui
 
 import androidx.fragment.app.Fragment
+import com.example.ishopping.R
 
 class LikeFragment : Fragment(R.layout.fragment_like) {
 }
